backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.26
botocore==1.36.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.0
Faker==37.0.2
fastapi==0.115.8
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
